pyproject.toml
[project]
name = "dify-llvm"
version = "0.1.0"
description = ""
authors = [
    {name = "tolatolatop",email = "tolatolatop@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydify @ git+https://github.com/jayscoder/pydify.git",
    "fastapi[standard] (>=0.118.0,<0.119.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "openai (>=2.1.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "ruamel-yaml (>=0.18.15,<0.19.0)"
]

[tool.poetry]
packages = [{include = "dify_llvm", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
